+++ dashboard/src/tests/msw/mocks/graphql/nhostGraphQLLink.ts
@@ -1,5 +1,5 @@
 import { graphql } from 'msw';
 
-const nhostGraphQLLink = graphql.link('http://localhost:1337/v1/graphql');
+const nhostGraphQLLink = graphql.link('https://local.graphql.nhost.run/v1');
 
 export default nhostGraphQLLink;