+++ examples/react-apollo/e2e/config.ts
@@ -1,0 +1,3 @@
+export const baseURL = 'http://localhost:3000'
+export const mailhogURL = 'http://localhost:8025'
+export const authBackendURL = 'https://local.auth.nhost.run'